slider组件支持0.x 步进，调整step类型为float

+++ erupt-annotation/src/main/java/xyz/erupt/annotation/sub_field/sub_edit/SliderType.java
@@ -12,7 +12,7 @@ public @interface SliderType {
     int min() default 0;
 
     @Comment("步进长度")
-    int step() default 1;
+    float step() default 1;
 
     @Comment("刻度标记")
     int[] markPoints() default {};
